Module re_build_info::crate_version

Modules

meta ğŸ”’

Macros

eat ğŸ”’

eat_token ğŸ”’

eat_u8 ğŸ”’

slice ğŸ”’

Slice s by some start and end bounds.

Structs

CrateVersion

The version of a Rerun crate.

Enums

Meta

Functions

const_u8_slice_util ğŸ”’

Helper function to slice slices in a const context. Instead of using this directly, use the slice macro.

equals ğŸ”’

maybe ğŸ”’

maybe_token ğŸ”’

split_at ğŸ”’