re_build_info::crate_version
Function const_u8_slice_util

const fn const_u8_slice_util(
    v: &[u8],
    start: Option<usize>,
    end: Option<usize>,
) -> &[u8] 

Helper function to slice slices in a const context. Instead of using this directly, use the slice macro.

This is equivalent to v[start..end].