re_build_info
Module crate_version

Modules

meta 🔒

Macros

slice 🔒

Slice s by some start and end bounds.

Structs

CrateVersion

The version of a Rerun crate.

Enums

Meta

Functions

const_u8_slice_util 🔒

Helper function to slice slices in a const context. Instead of using this directly, use the slice macro.

equals 🔒

split_at 🔒